+++ SharpBucket/V2/Pocos/TeamProfile.cs
@@ -5,6 +5,6 @@ namespace SharpBucket.V2.Pocos{
         public int? pagelen { get; set; }
         public List<Link> links { get; set; }
         public int? page { get; set; }
-        public int? size { get; set; }
+        public ulong? size { get; set; }
     }
 }
